:truck: Updating package name
