--- app/src/main/java/com/example/kikeou/MainActivityViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.kikeou
-
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModel
-
-class MainActivityViewModel : ViewModel() {
-    var currentFragment : MainActivity.Views = MainActivity.Views.Profile
-        private set
-
-    fun updateView(newView: MainActivity.Views) {
-        currentFragment = newView
-    }
-}
