--- app/src/main/java/com/example/kikeou/room/dao/ContactDao.kt
@@ -1,20 +1,0 @@
-package com.example.kikeou.room.dao
-
-import androidx.room.*
-import com.example.kikeou.room.models.Contact
-import kotlinx.coroutines.flow.Flow
-
-@Dao
-interface ContactDao {
-    @Query("SELECT * FROM contact")
-    fun getAll(): Flow<Contact>
-
-    @Insert
-    fun insert(contact: Contact)
-
-    @Delete
-    fun delete(contact: Contact)
-
-    @Update
-    fun update(contact: Contact)
-}
