+++ app/src/main/java/fr/enssat/kikeou/pepin_plestan_plusquellec/room/models/Localisation.kt
@@ -1,0 +1,15 @@
+package fr.enssat.kikeou.pepin_plestan_plusquellec.room.models
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+@Entity
+data class Localisation(
+    @Transient
+    @ColumnInfo(name="id") @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @ColumnInfo(name="day") val day: Int = 1,
+    @ColumnInfo(name="value") var value: String
+)
